pyproject.toml
[project]
name = "djangoormlearn"
version = "0.1.0"
description = ""
authors = [
    {name = "Vinandydy",email = "vinandy4718@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "faker (>=37.6.0,<38.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "django-stubs (>=5.2.7,<6.0.0)",
    "djangorestframework-stubs (>=3.16.5,<4.0.0)",
    "ruff (>=0.14.3,<0.15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
